Ci fixes (#141)
* ci: enable renovate for angular snapshots

With this change we enable renvoate to update the angular packages sha for snapshots and we group all these packages so that only 1 PR is created for all Angular packages.

We also enable scheduling, so that we don't get flooded by PRs when Angular caretaker merges the PRS. We will only run renovate during OOO hours.

* ci: use strict SHA for angular packages

With this change we create a `package.json` with Angular snapshot SHA. This will allow renovate to update these so that when we we create a new project the correct SHA are used.

For the build we use `npm ci` instead of `npm  i` so that we use frozen lock files

And also we disable `progress` and `sourceMap` which should help reduce the build times slightly.

* build: update lock files and also tsconfig to latest to avoid unused file warnings

Closes #136

* ci: use npm i instead of npm ci

We should use npm ci instead of npm i, so that we use frozen lock files. However currently the lock files are outdated and we need to update them.