Remove `@nguniversal/*` projects (#388)
Previously, the `@nguniversal/*` packages were imported into standard
`@angular/cli` apps. However, these packages are not supposed to be used
in Angular apps (on the client) and thus do not need to be processed by
`ngcc`. They contain code that must be run on the server (with access to
Node.js APIs) and thus the corresponding projects were failing during
build, due to missing built-in Node.js modules (e.g. `fs`).

We should definitely test how ngUniversal apps work with Ivy (and
`ngcc`), but this is not the right way. Removing the packages for now.